handle tool calls auto submit
